+++ CHANGELOG.md
@@ -1,4 +1,12 @@
-v1.5.3 (2018-01-
+v1.6.0 (2018-??-??)
+
+- Added support for serving image tiles (`.imtiles`)
+- Added support for serving image tile fragments, i.e., fragments from tile sets of filetype `imtiles`.
+- Added support for serving OpenStreetMap image tile fragments, i.e., fragments from tile sets of filetype `osm-image`.
+- Added support for serving "plain" 2d annotations (filetype `2dannodb`; datatype `2d-rectangle-domains`)
+- Added support for serving GeoJSON (filetype `geodb`; datatype `geo-json`)
+
+v1.5.3 (2018-??-??)
 
 - Refactored the chromsizes code to be more modular